+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `size-limit`
 * Remove `autoResolve`
 * Rewrite with `sham-ui-data-storage`
+* Rewrite `link-to` component with `href-to` directive
 
 ## 3.0.1
 * Fix build